requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
sqlalchemy==2.0.22
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
python-multipart>=0.0.9
alembic==1.12.0
requests==2.31.0
numpy==1.26.0
pandas==2.1.1
scikit-learn==1.3.1
tensorflow==2.14.0
python-dotenv==1.0.0
httpx==0.25.0
gunicorn==21.2.0
pytest==7.4.2
underthesea==6.7.0
langdetect==1.0.9
pyvi==0.1.1
gradio>=4.44.1
huggingface-hub==0.19.3
pydantic-settings>=2.0.0
scikit-learn==1.3.1
safetensors==0.4.0
transformers>=4.30.0
pydantic[email]>=2.0.0
email-validator>=2.0.0
redis>=5.0.0
hiredis>=2.2.0
prometheus-client>=0.19.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
httpx>=0.25.0